Hello World

+++ app/Http/Controllers/ProfileController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 use Inertia\Response;
+use Silber\Bouncer\BouncerFacade;
 
 class ProfileController extends Controller
 {
@@ -18,10 +19,17 @@ class ProfileController extends Controller
      */
     public function edit(Request $request): Response
     {
-        return Inertia::render('Profile/Edit', [
-            'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
-            'status' => session('status'),
-        ]);
+        if(Auth::user()->cannot('edit-profile'))
+        {
+            abort(403, 'Unauthorized access.');
+        }
+        else{
+            return Inertia::render('Profile/Edit', [
+                'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
+                'status' => session('status'),
+                'isAdmin' => BouncerFacade::is(Auth::user())->an('admin'),
+            ]);
+        }
     }
 
     /**
@@ -29,15 +37,21 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
-
-        if ($request->user()->isDirty('email')) {
-            $request->user()->email_verified_at = null;
+        if(Auth::user()->cannot('edit-profile'))
+        {
+            abort(403, 'Unauthorized access.');
         }
+        else{
+            $request->user()->fill($request->validated());
 
-        $request->user()->save();
+            if ($request->user()->isDirty('email')) {
+                $request->user()->email_verified_at = null;
+            }
 
-        return Redirect::route('profile.edit');
+            $request->user()->save();
+
+            return Redirect::route('profile.edit');
+        }
     }
 
     /**
@@ -45,19 +59,25 @@ class ProfileController extends Controller
      */
     public function destroy(Request $request): RedirectResponse
     {
-        $request->validate([
-            'password' => ['required', 'current_password'],
-        ]);
+        if(Auth::user()->cannot('edit-profile'))
+        {
+            abort(403, 'Unauthorized access.');
+        }
+        else{
+            $request->validate([
+                'password' => ['required', 'current_password'],
+            ]);
 
-        $user = $request->user();
+            $user = $request->user();
 
-        Auth::logout();
+            Auth::logout();
 
-        $user->delete();
+            $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
 
-        return Redirect::to('/');
+            return Redirect::to('/');
+        }
     }
 }
